added a utility checker/error handling function
The function will be used throught the program to check if the value entered
by the user is of a useable type for that specific command.

If the input value is appropiate for that command, the value will be converted
to the desired type, if not, the error is caught, a msg displayed, and the user
is returned to the main loop to try again.