updated commands.py with a pinch more of doc string content and fixed minor flaw
Feeling a little sleepy and unmotivated today, so i just added some better doc string
content to some functions and fixed a minor flaw in the 'addBid' function. The function
previously would allow the user the enter a word or something not useable as an integer
for the bidder id # which would confuse other functions, like the 'remBid' function for ex.

big oops

So thats better now lol